backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
asyncpg==0.29.0
alembic==1.13.1

# Redis
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6

# MinIO
minio==7.2.3

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# WebSockets
websockets==12.0
python-socketio[asgi]==5.11.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Utils
python-dotenv==1.0.0
aiofiles==23.2.1

# Scheduling (Phase 4.2)
apscheduler==3.10.4

# Test dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
aiosqlite>=0.19.0
